finished

+++ app/src/main/java/com/example/kotlinfirestore/adapter/ListAdapter.kt
@@ -1,15 +1,24 @@
 package com.example.kotlinfirestore.adapter
 
+import android.app.AlertDialog
 import android.content.Context
+import android.content.DialogInterface
+import android.content.Intent
+import android.graphics.Color
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
+import android.widget.LinearLayout
 import android.widget.TextView
+import android.widget.Toast
+import androidx.core.content.ContextCompat.startActivity
 import androidx.recyclerview.widget.RecyclerView
+import com.example.kotlinfirestore.PersonActivity
 import com.example.kotlinfirestore.R
 import com.example.kotlinfirestore.db.DatabaseHandler
 import com.example.kotlinfirestore.model.Person
+
 
 class ListAdapter (peopleList: List<Person>, internal var ctx: Context, private val callbacks: (Int) -> Unit): RecyclerView.Adapter<ListAdapter.ViewHolder>(){
     internal var peopleList: List<Person> = ArrayList()
@@ -27,10 +36,28 @@ class ListAdapter (peopleList: List<Person>, internal var ctx: Context, private 
         val person = peopleList[position]
 
         holder.name.text = person.name
+        if(position % 2 == 0) holder.parentLayout.setBackgroundColor(Color.GRAY)
+        else holder.parentLayout.setBackgroundColor(Color.WHITE)
+
+        holder.name.setOnClickListener {
+            val intent = Intent(ctx, PersonActivity::class.java)
+            intent.putExtra("edit", true)
+            intent.putExtra("person", person as java.io.Serializable)
+            ctx.startActivity(intent)
+        }
+
         holder.btn.setOnClickListener {
-            val dbHandler = DatabaseHandler(ctx)
-            dbHandler.deletePerson(person.id!!)
-            callbacks(position)
+            AlertDialog.Builder(ctx)
+                .setTitle("Deletar registro")
+                .setMessage("Você tem certeza que deseja deletar esse registro?")
+                .setIcon(android.R.drawable.ic_dialog_alert)
+                .setPositiveButton("Sim"
+                ) { dialog, whichButton ->
+                    callbacks(position)
+                    val dbHandler = DatabaseHandler(ctx)
+                    dbHandler.deletePerson(person.id!!)
+                }
+                .setNegativeButton("Não", null).show()
         }
     }
 
@@ -38,10 +65,9 @@ class ListAdapter (peopleList: List<Person>, internal var ctx: Context, private 
         return peopleList.size
     }
 
-
-
     inner class ViewHolder(view: View): RecyclerView.ViewHolder(view) {
         var name: TextView = view.findViewById(R.id.tvAdpName)
         var btn: Button = view.findViewById(R.id.btnAdpDel)
+        var parentLayout: LinearLayout = view.findViewById(R.id.parent_layout)
     }
 }
